chore: CoordinateVO 공통 경로로 이동
