+++ eatngo-infra/src/main/kotlin/com/eatngo/store/rdb/entity/StoreJpaEntity.kt
@@ -2,6 +2,7 @@ package com.eatngo.store.rdb.entity
 
 import com.eatngo.common.BaseJpaEntity
 import com.eatngo.common.constant.StoreEnum
+import com.eatngo.common.type.CoordinateVO
 import com.eatngo.constants.DELETED_FILTER
 import com.eatngo.store.domain.*
 import com.eatngo.store.vo.*
